pyproject.toml
[tool.poetry]
name = "Synapse"
version = "0.1.0"
description = ""
authors = ["IK <ivandkyulev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
tabulate = "^0.8.10"
selenium = "^4.4.3"
webdriver-manager = "^3.8.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
